Feature/GC-1142 add linters

Approved-by: Dmitry Danilov

+++ logger/logger.go
@@ -16,10 +16,13 @@ import (
 var Logger FtLogger
 
 func init() {
-	godotenv.Load(".env")
+	err := godotenv.Load(".env")
+	if err != nil {
+		logrus.Warnln("unable to read .env, falling back to os variables")
+	}
 	Logger.logger = logrus.New()
 
-	err := Logger.SetLevel(os.Getenv("LOG_LEVEL"))
+	err = Logger.SetLevel(os.Getenv("LOG_LEVEL"))
 	if err != nil {
 		logrus.Errorf("set level: %s", err)
 	}
@@ -32,7 +35,6 @@ func init() {
 	if err != nil {
 		logrus.Errorf("set log file: %s", err)
 	}
-
 }
 
 type FtLogger struct {
@@ -52,21 +54,21 @@ func (l *FtLogger) isWriteToFile() bool {
 func (l *FtLogger) prepLogFile() error {
 	cacheDir, err := os.UserCacheDir()
 	if err != nil {
-		return fmt.Errorf("get user cache dir: %s", err)
+		return fmt.Errorf("get user cache dir: %w", err)
 	}
 
 	logFileDir := filepath.Dir(l.logFilePath)
 	logFileDirPath := filepath.Join(cacheDir, logFileDir)
-	err = os.MkdirAll(logFileDirPath, 0755)
+	err = os.MkdirAll(logFileDirPath, 0o755)
 	if err != nil {
-		return fmt.Errorf("make \"%s\" dir: %s", logFileDirPath, err)
+		return fmt.Errorf("make \"%s\" dir: %w", logFileDirPath, err)
 	}
 
 	logFilePath := filepath.Join(cacheDir, l.logFilePath)
 
-	l.logFile, err = os.OpenFile(logFilePath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0755)
+	l.logFile, err = os.OpenFile(logFilePath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0o755)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to open the file: %w", err)
 	}
 
 	return nil
@@ -95,7 +97,7 @@ func (l *FtLogger) SetLevel(level string) error {
 	}
 	lvl, err := logrus.ParseLevel(level)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to parse level: %w", err)
 	}
 
 	l.logger.SetLevel(lvl)
@@ -123,7 +125,10 @@ func (l *FtLogger) SetLoggerOutput(path string) error {
 }
 
 func (l *FtLogger) Close() error {
-	l.SendLogsToController()
+	err := l.SendLogsToController()
+	if err != nil {
+		logrus.Errorln(err)
+	}
 	if l.isWriteToFile() {
 		err := l.logFile.Close()
 		if err != nil {
@@ -132,31 +137,28 @@ func (l *FtLogger) Close() error {
 
 			return errors.New(msg)
 		}
-
 	}
 
 	return nil
 }
 
 func (l *FtLogger) SendLogsToController() error {
-	if l.isWriteToFile() {
-		if l.sdr != nil {
-			err := l.sendLogs()
-			if err != nil {
-				msg := fmt.Sprintf("send logs to controller: %s", err)
-				logrus.Error(msg)
+	if l.isWriteToFile() && l.sdr != nil {
+		err := l.sendLogs()
+		if err != nil {
+			msg := fmt.Sprintf("send logs to controller: %s", err)
+			logrus.Error(msg)
 
-				return errors.New(msg)
-			}
-			cacheDir, err := os.UserCacheDir()
-			if err != nil {
-				return fmt.Errorf("get user cache dir: %s", err)
-			}
-			path := filepath.Join(cacheDir, l.logFilePath)
-			err = os.Truncate(path, 0)
-			if err != nil {
-				logrus.Error("truncating log file: ", err)
-			}
+			return errors.New(msg)
+		}
+		cacheDir, err := os.UserCacheDir()
+		if err != nil {
+			return fmt.Errorf("get user cache dir: %w", err)
+		}
+		path := filepath.Join(cacheDir, l.logFilePath)
+		err = os.Truncate(path, 0)
+		if err != nil {
+			logrus.Error("truncating log file: ", err)
 		}
 	}
 	return nil
@@ -165,12 +167,12 @@ func (l *FtLogger) SendLogsToController() error {
 func (l *FtLogger) sendLogs() error {
 	cacheDir, err := os.UserCacheDir()
 	if err != nil {
-		return fmt.Errorf("get user cache dir: %v", err)
+		return fmt.Errorf("get user cache dir: %w", err)
 	}
 	logFilePath := filepath.Join(cacheDir, l.logFilePath)
 	content, err := ioutil.ReadFile(logFilePath)
 	if err != nil {
-		return fmt.Errorf("error in reading log file: %v", err)
+		return fmt.Errorf("error in reading log file: %w", err)
 	}
 	reqData := request{
 		MachineID:  l.machineID(),
@@ -180,7 +182,7 @@ func (l *FtLogger) sendLogs() error {
 
 	requestBody, err := json.Marshal(reqData)
 	if err != nil {
-		return fmt.Errorf("request body: %v", err)
+		return fmt.Errorf("request body: %w", err)
 	}
 
 	return l.sdr.send(requestBody)
